fix(kurtosis/config): fix kurtosis config to use teku instead of nimbus. (#1852)
## Description

Fixes the kurtosis configuration to use `teku` instead of `nimbus` for
L1 consensus client. The latest version of `nimbus` seems to have broken
the `optimism-package`.

Also sets the `cl_image` to point to a local image of the node by
default. This will improve DX until we reach more stability